[MOV] automatic_sum: move the store into the correct folder
Task: 6394129
